+++ packages/jwt-verifier/lib.js
@@ -13,10 +13,52 @@
 const jwksClient = require('jwks-rsa');
 const nJwt = require('njwt');
 
-const {
-  assertIssuer,
-  assertClientId
-} = require('@okta/configuration-validation');
+class ConfigurationValidationError extends Error {}
+
+const findDomainURL = 'https://bit.ly/finding-okta-domain';
+const findAppCredentialsURL = 'https://bit.ly/finding-okta-app-credentials';
+
+const assertIssuer = (issuer, testing = {}) => {
+  const isHttps = new RegExp('^https://');
+  const hasDomainAdmin = /-admin.(okta|oktapreview|okta-emea).com/;
+  const copyMessage = 'You can copy your domain from the Okta Developer ' +
+    'Console. Follow these instructions to find it: ' + findDomainURL;
+
+  if (testing.disableHttpsCheck) {
+    const httpsWarning = 'Warning: HTTPS check is disabled. ' +
+      'This allows for insecure configurations and is NOT recommended for production use.';
+    /* eslint-disable-next-line no-console */
+    console.warn(httpsWarning);
+  }
+
+  if (!issuer) {
+    throw new ConfigurationValidationError('Your Okta URL is missing. ' + copyMessage);
+  } else if (!testing.disableHttpsCheck && !issuer.match(isHttps)) {
+    throw new ConfigurationValidationError(
+      'Your Okta URL must start with https. ' +
+      `Current value: ${issuer}. ${copyMessage}`
+    );
+  } else if (issuer.match(/{yourOktaDomain}/)) {
+    throw new ConfigurationValidationError('Replace {yourOktaDomain} with your Okta domain. ' + copyMessage);
+  } else if (issuer.match(hasDomainAdmin)) {
+    throw new ConfigurationValidationError(
+      'Your Okta domain should not contain -admin. ' +
+      `Current value: ${issuer}. ${copyMessage}`
+    );
+  }
+};
+
+const assertClientId = (clientId) => {
+  const copyCredentialsMessage = 'You can copy it from the Okta Developer Console ' +
+    'in the details for the Application you created. ' +
+    `Follow these instructions to find it: ${findAppCredentialsURL}`;
+
+  if (!clientId) {
+    throw new ConfigurationValidationError('Your client ID is missing. ' + copyCredentialsMessage);
+  } else if (clientId.match(/{clientId}/)) {
+    throw new ConfigurationValidationError('Replace {clientId} with the client ID of your Application. ' + copyCredentialsMessage);
+  }
+};
 
 class AssertedClaimsVerifier {
   constructor() {
@@ -41,18 +83,18 @@ class AssertedClaimsVerifier {
 
   isValidOperator(operator) {
     // may support more operators in the future
-    return !operator || operator === 'includes'
+    return !operator || operator === 'includes';
   }
 
   checkAssertions(op, claim, expectedValue, actualValue) {
     if (!op && actualValue !== expectedValue) {
       this.errors.push(`claim '${claim}' value '${actualValue}' does not match expected value '${expectedValue}'`);
     } else if (op === 'includes' && Array.isArray(expectedValue)) {
-      expectedValue.forEach(value => {
+      expectedValue.forEach((value) => {
         if (!actualValue || !actualValue.includes(value)) {
           this.errors.push(`claim '${claim}' value '${actualValue}' does not include expected value '${value}'`);
         }
-      })
+      });
     } else if (op === 'includes' && (!actualValue || !actualValue.includes(expectedValue))) {
       this.errors.push(`claim '${claim}' value '${actualValue}' does not include expected value '${expectedValue}'`);
     }
@@ -61,14 +103,14 @@ class AssertedClaimsVerifier {
 
 function verifyAssertedClaims(verifier, claims) {
   const assertedClaimsVerifier = new AssertedClaimsVerifier();
-  for (let [claimName, expectedValue] of Object.entries(verifier.claimsToAssert)) {
+  for (const [claimName, expectedValue] of Object.entries(verifier.claimsToAssert)) {
     const operator = assertedClaimsVerifier.extractOperator(claimName);
     if (!assertedClaimsVerifier.isValidOperator(operator)) {
       throw new Error(`operator: '${operator}' invalid. Supported operators: 'includes'.`);
     }
     const claim = assertedClaimsVerifier.extractClaim(claimName);
     const actualValue = claims[claim];
-    assertedClaimsVerifier.checkAssertions(operator, claim, expectedValue, actualValue)
+    assertedClaimsVerifier.checkAssertions(operator, claim, expectedValue, actualValue);
   }
   if (assertedClaimsVerifier.errors.length) {
     throw new Error(assertedClaimsVerifier.errors.join(', '));
@@ -76,67 +118,72 @@ function verifyAssertedClaims(verifier, claims) {
 }
 
 function verifyAudience(expected, aud) {
-  if( !expected ) {
+  if (!expected) {
     throw new Error('expected audience is required');
   }
 
-  if ( Array.isArray(expected) && !expected.includes(aud) ) {
-    throw new Error(`audience claim ${aud} does not match one of the expected audiences: ${ expected.join(', ') }`);
+  if (Array.isArray(expected) && !expected.includes(aud)) {
+    throw new Error(`audience claim ${aud} does not match one of the expected audiences: ${expected.join(', ')}`);
   }
 
-  if ( !Array.isArray(expected) && aud !== expected ) {
+  if (!Array.isArray(expected) && aud !== expected) {
     throw new Error(`audience claim ${aud} does not match expected audience: ${expected}`);
   }
 }
 
 function verifyClientId(expected, aud) {
-  if( !expected ) {
+  if (!expected) {
     throw new Error('expected client id is required');
   }
 
   assertClientId(expected);
 
-  if ( aud !== expected ) {
+  if (aud !== expected) {
     throw new Error(`audience claim ${aud} does not match expected client id: ${expected}`);
   }
 }
 
 function verifyIssuer(expected, issuer) {
-  if( issuer !== expected ) {
+  if (issuer !== expected) {
     throw new Error(`issuer ${issuer} does not match expected issuer: ${expected}`);
   }
 }
 
 function verifyNonce(expected, nonce) {
-  if( nonce && !expected ) {
+  if (nonce && !expected) {
     throw new Error('expected nonce is required');
   }
   if (!nonce && expected) {
     throw new Error(`nonce claim is missing but expected: ${expected}`);
   }
-  if( nonce && expected && nonce !== expected ) {
+  if (nonce && expected && nonce !== expected) {
     throw new Error(`nonce claim ${nonce} does not match expected nonce: ${expected}`);
   }
+}
+
+function getJwksUri(options) {
+  const baseUrl = options.issuer?.indexOf('/oauth2') > 0 ? options.issuer : options.issuer + '/oauth2'; // Handle org AS
+  return options.jwksUri ? options.jwksUri : baseUrl + '/v1/keys';
 }
 
 class OktaJwtVerifier {
   constructor(options = {}) {
     // Assert configuration options exist and are well-formed (not necessarily correct!)
     assertIssuer(options.issuer, options.testing);
-    if( options.clientId ) {
+    if (options.clientId) {
       assertClientId(options.clientId);
     }
 
     this.claimsToAssert = options.assertClaims || {};
     this.issuer = options.issuer;
-    const baseUrl = issuer?.indexOf('/oauth2') > 0 ? this.issuer : this.issuer + '/oauth2'; // Handle org AS
+    this.jwksUri = getJwksUri(options);
     this.jwksClient = jwksClient({
-      jwksUri: baseUrl + '/v1/keys',
+      jwksUri: this.jwksUri,
       cache: true,
       cacheMaxAge: options.cacheMaxAge || (60 * 60 * 1000),
       cacheMaxEntries: 3,
       jwksRequestsPerMinute: options.jwksRequestsPerMinute || 10,
-      rateLimit: true
+      rateLimit: true,
     });
     this.verifier = nJwt.createVerifier().setSigningAlgorithm('RS256').withKeyResolver((kid, cb) => {
       if (kid) {
@@ -144,7 +191,7 @@ class OktaJwtVerifier {
           cb(err, key && (key.publicKey || key.rsaPublicKey));
         });
       } else {
-        cb("No KID specified", null);
+        cb('No KID specified', null);
       }
     });
   }