+++ packages/oidc-middleware/CHANGELOG.md
@@ -1,3 +1,15 @@
+# 4.2.0
+
+### Bug Fixes
+
+- [#1020](https://github.com/okta/okta-oidc-js/pull/1020) Fixes issue with UUID returning null
+
+# 4.1.0
+
+### Features
+
+- [#989](https://github.com/okta/okta-oidc-js/pull/989) adds callback to allow custom error handling
+  
 # 4.0.3
 
 ### Bug Fixes